Move subsidiary serializer to serializers file
